HV_Control_SW/requirements.txt
pyqt5
pyqtgraph
pyserial
pyvisa
pyUSB
numpy
caen-libs
